+++ src/main/java/com/coolerpromc/archerythings/model/quiver/QuiverModel.java
@@ -9,7 +9,7 @@ import net.minecraft.resources.ResourceLocation;
 
 import java.util.function.Function;
 
-public class QuiverModel extends Model {
+public class QuiverModel<S> extends Model<S> {
     private final ModelPart root;
     private final ModelPart quiver;