+++ src/main/java/com/coolerpromc/archerythings/util/LivingEntityRenderStateAccessor.java
@@ -1,0 +1,8 @@
+package com.coolerpromc.archerythings.util;
+
+import net.minecraft.world.item.ItemStack;
+
+public interface LivingEntityRenderStateAccessor {
+    void archerythings$setQuiverStack(ItemStack quiverStack);
+    ItemStack archerythings$getQuiverStack();
+}